docker/wis2bridge/requirements.txt
paho-mqtt
awsiotsdk
boto3